Revert "Specification by Example"
