--- integration-test/src/test/java/org/jboss/pnc/integration/util/TestUtils.java
@@ -1,47 +1,0 @@
-package org.jboss.pnc.integration.util;
-
-import org.hamcrest.CustomMatcher;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.lang.invoke.MethodHandles;
-import java.util.function.Consumer;
-import java.util.stream.Stream;
-
-import static com.jayway.restassured.path.json.JsonPath.from;
-
-
-public class TestUtils {
-
-    public static final Logger logger = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
-
-    public static CustomMatcher<String> containsJsonAttribute(String jsonAttribute, Consumer<String>... actionWhenMatches) {
-        return new CustomMatcher<String>("matchesJson") {
-            @Override
-            public boolean matches(Object o) {
-                String rawJson = String.valueOf(o).intern();
-                logger.debug("Evaluating raw JSON: " + rawJson);
-                Object value = from(rawJson).get(jsonAttribute);
-                logger.debug("Got value from JSon: " + value);
-                if(value != null) {
-                    if(actionWhenMatches != null) {
-                        Stream.of(actionWhenMatches).forEach(action -> action.accept(String.valueOf(value)));
-                    }
-                    return true;
-                }
-                return false;
-            }
-        };
-    }
-
-    /**
-     * Gets Test Http Port.
-     *
-     * @return the http port for REST end points, default to 8080.
-     */
-    public static int getHttpPort() {
-        int defaultHttpPort = Integer.getInteger("jboss.http.port", 8080);
-        int offset = Integer.getInteger("jboss.port.offset", 0);
-        return defaultHttpPort + offset;
-    }
-}
